requirements.txt
flask>=3.1.0
psutil >= 7.0.0
celery[redis] >= 5.5.3
flask-wtf >= 1.2.2
markdown >= 3.10
pyotp >= 2.9.0
qrcode >= 8.2
pillow >= 11.2.1
bcrypt >= 5.0.0
flask-session >= 0.8.0
redis >= 5.2
pySMART >= 1.4.2
flask-babel >= 4.0.0
fastapi >= 0.128.0
uvicorn >= 0.40.0
pyjwt >= 2.10.1
python-multipart >= 0.0.22
requests >= 2.32.3